Pulled out the PositionOnMonitorSelector which was a rather large chunk of the xaml in the PreferencesView
